lua5.4: (WIP) moving 95% of Lua related code to sjasm/lua_sjasm.cpp
it's new "bindings/wrapper" source file.

No other source file should operate with Lua C API directly, except
io_err.cpp using lua API to figure details about error (and feels like
this would be difficult to decouple).
